# argus/docs/swagger.yaml
basePath: /
definitions:
...
  argus_app_internal_model.Department:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      leader:
        type: string
      name:
        type: string
      parentId:
        description: 0=根
        type: integer
      phone:
        type: string
      sort:
        type: integer
      status:
        description: 1 启用 / 0 禁用；新建默认值由 service 明确填充
        type: integer
      updatedAt:
        type: string
    type: object
  argus_app_internal_model.DepartmentTreeNode:
    properties:
      children:
        items:
          $ref: '#/definitions/argus_app_internal_model.DepartmentTreeNode'
        type: array
      createdAt:
        type: string
      id:
        type: integer
      leader:
        type: string
      name:
        type: string
      parentId:
        description: 0=根
        type: integer
      phone:
        type: string
      sort:
        type: integer
      status:
        description: 1 启用 / 0 禁用；新建默认值由 service 明确填充
        type: integer
      updatedAt:
        type: string
    type: object
...
  argus_app_internal_model.Menu:
    properties:
      affix:
        type: boolean
      component:
        description: 视图相对路径；按钮级为空
        type: string
      createdAt:
        type: string
      homePath:
        type: string
      icon:
        type: string
      id:
        type: integer
      keepAlive:
        type: boolean
      name:
        description: catalog/menu 为 ASCII 路由标识符；button 为 i18n key（如 system.user.addUser）
        type: string
      parentId:
        description: 0=根
        type: integer
      path:
        type: string
      permission:
        description: catalog/menu 存模块码，button 存动作码
        type: string
      sort:
        type: integer
      status:
        description: 类型由 gorm 按驱动映射（决策 18）
        type: integer
      title:
        description: i18n key，如 routes.system.user（决策 17）
        type: string
      type:
        description: catalog / menu / button
        type: string
      updatedAt:
        type: string
    type: object
  argus_app_internal_model.MenuTreeNode:
    properties:
      affix:
        type: boolean
      children:
        items:
          $ref: '#/definitions/argus_app_internal_model.MenuTreeNode'
        type: array
      component:
        description: 视图相对路径；按钮级为空
        type: string
      createdAt:
        type: string
      homePath:
        type: string
      icon:
        type: string
      id:
        type: integer
      keepAlive:
        type: boolean
      name:
        description: catalog/menu 为 ASCII 路由标识符；button 为 i18n key（如 system.user.addUser）
        type: string
      parentId:
        description: 0=根
        type: integer
      path:
        type: string
      permission:
        description: catalog/menu 存模块码，button 存动作码
        type: string
      sort:
        type: integer
      status:
        description: 类型由 gorm 按驱动映射（决策 18）
        type: integer
      title:
        description: i18n key，如 routes.system.user（决策 17）
        type: string
      type:
        description: catalog / menu / button
        type: string
      updatedAt:
        type: string
    type: object
  argus_app_internal_model.OperationLog:
    properties:
      action:
        type: string
      body:
        description: 已脱敏 JSON
        type: string
      createdAt:
        description: CreatedAt 覆盖 TimeFields：操作日志按时间范围查询，需要索引（设计 §2）。
        type: string
      durationMs:
        type: integer
      id:
        type: integer
      ip:
        type: string
      method:
        type: string
      module:
        type: string
      path:
        type: string
      query:
        type: string
      statusCode:
        type: integer
      updatedAt:
        type: string
      userAgent:
        type: string
      userId:
        description: 登录失败等场景可为空
        type: integer
      username:
        type: string
    type: object
...
  argus_app_internal_pkg_response.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  argus_app_internal_service.LogPageResult:
    properties:
      items:
        items:
          $ref: '#/definitions/argus_app_internal_model.OperationLog'
        type: array
      total:
        type: integer
    type: object
...
  argus_app_internal_service.SaveDeptInput:
    properties:
      leader:
        type: string
      name:
        type: string
      parentId:
        description: 0=根
        type: integer
      phone:
        type: string
      sort:
        type: integer
      status:
        description: 省略时新建默认启用；编辑省略表示不变
        type: integer
    required:
    - name
    type: object
  argus_app_internal_service.SaveMenuInput:
    properties:
      affix:
        type: boolean
      component:
        type: string
      homePath:
        type: string
      icon:
        type: string
      keepAlive:
        type: boolean
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      permission:
        type: string
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        enum:
        - catalog
        - menu
        - button
        type: string
    required:
    - name
    - type
    type: object
...
  argus_app_internal_service.UploadedFile:
    properties:
      contentType:
        type: string
      key:
        type: string
      name:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
...
  argus_app_internal_service.VbenRouteMeta:
    properties:
      affixTab:
        type: boolean
      icon:
        type: string
      keepAlive:
        type: boolean
      order:
        type: integer
      title:
        type: string
    type: object
  argus_app_internal_service.VbenRouteRecord:
    properties:
      children:
        items:
          $ref: '#/definitions/argus_app_internal_service.VbenRouteRecord'
        type: array
      component:
        type: string
      id:
        type: integer
      meta:
        $ref: '#/definitions/argus_app_internal_service.VbenRouteMeta'
      name:
        type: string
      path:
        type: string
      pid:
        type: integer
      type:
        type: string
    type: object
...
  internal_api.DeptResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_model.Department'
      message:
        type: string
    type: object
  internal_api.DeptTreeResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/argus_app_internal_model.DepartmentTreeNode'
        type: array
      message:
        type: string
    type: object
  internal_api.FileUploadResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_service.UploadedFile'
      message:
        type: string
    type: object
...
  internal_api.LogPageResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_service.LogPageResult'
      message:
        type: string
    type: object
  internal_api.LogResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_model.OperationLog'
      message:
        type: string
    type: object
...
  internal_api.MenuResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_model.Menu'
      message:
        type: string
    type: object
  internal_api.MenuTreeResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/argus_app_internal_model.MenuTreeNode'
        type: array
      message:
        type: string
    type: object
  internal_api.NilResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_api.UserMenuTreeResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/argus_app_internal_service.VbenRouteRecord'
        type: array
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: argus 后端 API 接口文档
  title: argus API
  version: "1.0"
paths:
...
  /api/dept:
    post:
      consumes:
      - application/json
      description: 新增部门节点
      parameters:
      - description: 创建部门参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveDeptInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的部门信息
          schema:
            $ref: '#/definitions/internal_api.DeptResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 创建部门
      tags:
      - 部门模块
  /api/dept/{id}:
    delete:
      description: 删除指定 ID 的部门（有子部门或包含用户时不可删除）
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 删除部门
      tags:
      - 部门模块
    put:
      consumes:
      - application/json
      description: 更新指定 ID 的部门信息
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新部门参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveDeptInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的部门信息
          schema:
            $ref: '#/definitions/internal_api.DeptResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 更新部门
      tags:
      - 部门模块
  /api/dept/tree:
    get:
      description: 查询系统所有部门并构建为树形结构
      produces:
      - application/json
      responses:
        "200":
          description: 部门树数据
          schema:
            $ref: '#/definitions/internal_api.DeptTreeResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取全量部门树
      tags:
      - 部门模块
  /api/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片或 PDF 文件，返回不需要认证即可访问的文件 URL
      parameters:
      - description: 上传文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            $ref: '#/definitions/internal_api.FileUploadResponse'
        "400":
          description: 文件参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "500":
          description: 存储失败
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件模块
  /api/menu:
    post:
      consumes:
      - application/json
      description: 新增菜单、目录或权限按钮节点
      parameters:
      - description: 创建菜单参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveMenuInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的菜单信息
          schema:
            $ref: '#/definitions/internal_api.MenuResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 创建菜单/按钮
      tags:
      - 菜单模块
  /api/menu/{id}:
    delete:
      description: 删除指定 ID 的菜单（有子菜单时不可删除）
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 删除菜单/按钮
      tags:
      - 菜单模块
    put:
      consumes:
      - application/json
      description: 更新指定 ID 的菜单节点信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新菜单参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveMenuInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的菜单信息
          schema:
            $ref: '#/definitions/internal_api.MenuResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 更新菜单/按钮
      tags:
      - 菜单模块
  /api/menu/all:
    get:
      description: 根据当前用户角色权限，生成用于前端动态挂载路由的菜单结构
      produces:
      - application/json
      responses:
        "200":
          description: 前端动态路由树
          schema:
            $ref: '#/definitions/internal_api.UserMenuTreeResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取当前用户可访问菜单树
      tags:
      - 菜单模块
  /api/menu/tree:
    get:
      description: 查询系统所有菜单/按钮并构建为树形结构（用于菜单管理页面）
      produces:
      - application/json
      responses:
        "200":
          description: 菜单树数据
          schema:
            $ref: '#/definitions/internal_api.MenuTreeResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取全量菜单树
      tags:
      - 菜单模块
  /api/oplog/{id}:
    get:
      description: 根据 ID 查询单条操作日志详情
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 操作日志详情
          schema:
            $ref: '#/definitions/internal_api.LogResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取操作日志详情
      tags:
      - 操作日志模块
  /api/oplog/page:
    get:
      description: 分页查询用户操作日志，支持用户名、模块名称、状态码以及时间范围筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        in: query
        name: pageSize
        type: integer
      - description: 操作用户名
        in: query
        name: username
        type: string
      - description: 功能模块
        in: query
        name: module
        type: string
      - description: HTTP响应状态码
        in: query
        name: statusCode
        type: integer
      - description: '开始时间 (RFC3339 格式: 2006-01-02T15:04:05Z07:00)'
        in: query
        name: startTime
        type: string
      - description: '结束时间 (RFC3339 格式: 2006-01-02T15:04:05Z07:00)'
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作日志分页数据
          schema:
            $ref: '#/definitions/internal_api.LogPageResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 分页获取操作日志列表
      tags:
      - 操作日志模块
...
securityDefinitions:
  BearerAuth:
    description: 在 Header 中传入 `Bearer <token>` 进行身份认证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"